psets-6/requirements.txt
Pillow==9.1.1
tabulate==0.8.9
